requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
watchfiles==0.21.0
python-dotenv==1.0.0
tqdm==4.66.1
aiohttp==3.8.6
aiofiles==23.2.1
pillow==10.1.0
requests==2.31.0 